// lib/database.types.ts
export interface Database {
// ...
  public: {
    Tables: {
// ...
      assignments: {
        Row: {
          assignment_id: string
          attachment: string | null
          course_id: string | null
          description: string | null
          due_date: string | null
          instructor_id: string | null
          link: string | null
          title: string | null
        }
        Insert: {
          assignment_id?: string
          attachment?: string | null
          course_id?: string | null
          description?: string | null
          due_date?: string | null
          instructor_id?: string | null
          link?: string | null
          title?: string | null
        }
        Update: {
          assignment_id?: string
          attachment?: string | null
          course_id?: string | null
          description?: string | null
          due_date?: string | null
          instructor_id?: string | null
          link?: string | null
          title?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "assignments_instructor_id_fkey"
            columns: ["instructor_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["profile_id"]
          },
          {
            foreignKeyName: "course_id_fk"
            columns: ["course_id"]
            isOneToOne: false
            referencedRelation: "courses"
            referencedColumns: ["course_id"]
          }
        ]
      }
      courses: {
        Row: {
          course_description: string | null
          course_id: string
          course_name: string
          instructor_id: string
          room: string | null
          section: string | null
          subject: string | null
        }
        Insert: {
          course_description?: string | null
          course_id?: string
          course_name: string
          instructor_id?: string
          room?: string | null
          section?: string | null
          subject?: string | null
        }
        Update: {
          course_description?: string | null
          course_id?: string
          course_name?: string
          instructor_id?: string
          room?: string | null
          section?: string | null
          subject?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "fk_instructor_id"
            columns: ["instructor_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["profile_id"]
          }
        ]
      }
// ...
      profiles: {
        Row: {
          avatar_url: string | null
          biography: string | null
          email: string
          full_name: string | null
          position: string | null
          profile_id: string
          program: string | null
          role: Database["public"]["Enums"]["role_type"]
          section: string | null
          university: string | null
          updated_at: string | null
          username: string | null
        }
        Insert: {
          avatar_url?: string | null
          biography?: string | null
          email: string
          full_name?: string | null
          position?: string | null
          profile_id: string
          program?: string | null
          role?: Database["public"]["Enums"]["role_type"]
          section?: string | null
          university?: string | null
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          avatar_url?: string | null
          biography?: string | null
          email?: string
          full_name?: string | null
          position?: string | null
          profile_id?: string
          program?: string | null
          role?: Database["public"]["Enums"]["role_type"]
          section?: string | null
          university?: string | null
          updated_at?: string | null
          username?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "profiles_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: true
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      get_username_from_email: {
        Args: {
          email_address: string
        }
        Returns: string
      }
    }
    Enums: {
      post_type: "assignment" | "announcement"
      role_type: "student" | "instructor"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (Database["public"]["Tables"] & Database["public"]["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (Database["public"]["Tables"] &
      Database["public"]["Views"])
  ? (Database["public"]["Tables"] &
      Database["public"]["Views"])[PublicTableNameOrOptions] extends {
      Row: infer R
    }
    ? R
    : never
  : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof Database["public"]["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof Database["public"]["Enums"]
  ? Database["public"]["Enums"][PublicEnumNameOrOptions]
  : never